Enable instructions unification
Like Intel's Macro OPs fusion, unify a conditinal instruction and
a branching instruction. It speeds up while loop 1.2x.

Speedup ratio: compare with the result of `ruby 2.4.0dev (2016-05-04 trunk 54911) [x86_64-l[34/1900]reater is better)
name    built-ruby
loop_whileloop       1.250
vm1_attr_ivar*       1.018
vm1_attr_ivar_set*           0.991
vm1_block*           1.004
vm1_const*           2.081
vm1_ensure*          0.829
vm1_float_simple*            0.980
vm1_gc_short_lived*          0.997
vm1_gc_short_with_complex_long*      1.004
vm1_gc_short_with_long*      1.004
vm1_gc_short_with_symbol*            1.009
vm1_gc_wb_ary*       0.972
vm1_gc_wb_ary_promoted*      0.971
vm1_gc_wb_obj*       0.942
vm1_gc_wb_obj_promoted*      0.974
vm1_ivar*            0.975
vm1_ivar_set*        0.956
vm1_length*          0.996
vm1_lvar_init*       1.227
vm1_lvar_set*        0.997
vm1_neq*             1.044
vm1_not*             0.990
vm1_rescue*          0.942
vm1_simplereturn*            0.990
vm1_swap*            0.942
vm1_yield*           1.004

+++ compile.c
@@ -2506,7 +2506,7 @@ iseq_insns_unification(rb_iseq_t *iseq, LINK_ANCHOR *anchor)
 		    LINK_ELEMENT *li = list->next;
 		    for (k = 2; k < unified[1]; k++) {
 			if (!IS_INSN(li) ||
-			    ((INSN *)li)->insn_id != unified[k]) {
+			    (int)((INSN *)li)->insn_id != unified[k]) {
 			    goto miss;
 			}
 			li = li->next;
